CLI module
Abstracting Command line related code into it's own module
This allows the main logic to be exposed to other libraries without
requiring files to exist
