+++ lib/mson2json.js
@@ -1,13 +1,6 @@
-(function (Drafter, bpCompiler, fs, Transaction) {
+(function (Drafter, bpCompiler, Transaction) {
 
     'use strict';
-
-    function loadMson(path, cb) {
-        fs.readFile(path, function (err, data) {
-            if (err) throw err;
-            return cb(data.toString());
-        });
-    }
 
     function parseBluprint(data, cb) {
         var drafter = new Drafter;
@@ -63,20 +56,17 @@
     }
 
     /**
-     * path - path to MSON blueprint
+     * msonData - mson data from anywhere
      * cb   - function (JSON data)
      */
-    module.exports = function main(path, cb) {
-        return loadMson(path, function (msonData) {
-            parseBluprint(msonData, function (parsedData) {
-                convertToJson(parsedData, cb);
-            });
+    module.exports = function main(msonData, cb) {
+        return parseBluprint(msonData, function (parsedData) {
+            convertToJson(parsedData, cb);
         });
     };
 
 })(
     require('drafter'),
     require('blueprint-transactions'),
-    require('fs'),
     require('./transaction')
 );
